Moved load image to image.py
