+++ instadam/app.py
@@ -32,8 +32,6 @@ def create_app(mode='development'):
     app.register_blueprint(error_handlers.bp)
     from . import image
     app.register_blueprint(image.bp)
-    from . import project
-    app.register_blueprint(project.bp)
 
     if not os.path.isdir(app.config['STATIC_STORAGE_DIR']):
         os.mkdir(app.config['STATIC_STORAGE_DIR'])
